Move ItemCell in a dedicated file
