+++ Sources/Castor/UserInterface/Player/ItemCell.swift
@@ -1,0 +1,38 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import SwiftUI
+
+struct ItemCell: View {
+    @ObservedObject var item: CastPlayerItem
+
+    var body: some View {
+        HStack(spacing: 30) {
+            artworkImage()
+            Text(title)
+                .redacted(!item.isFetched)
+            Spacer()
+            disclosureImage()
+        }
+        .onAppear(perform: item.fetch)
+    }
+
+    private var title: String {
+        guard item.isFetched else { return .placeholder(length: .random(in: 20...30)) }
+        return item.metadata?.title ?? "Untitled"
+    }
+
+    private func artworkImage() -> some View {
+        ArtworkImage(url: item.metadata?.imageUrl(matching: .init(type: .custom, size: .init(width: 45, height: 45))))
+            .shadow(color: .primary.opacity(0.15), radius: 6, y: 3)
+            .frame(height: 45)
+    }
+
+    private func disclosureImage() -> some View {
+        Image(systemName: "line.3.horizontal")
+            .foregroundStyle(.secondary)
+    }
+}
